+++ infra/src/CloudRunApp.ts
@@ -3,10 +3,31 @@ import {BaseConstruct} from './baseConstruct';
 import {Secret} from './constructs/secretManager';
 import {ServiceAccount} from './constructs/serviceAccount';
 import {CloudRun} from './constructs/cloudRun';
+import {CloudBuild} from './constructs/CloudBuild';
+import {ArRepository} from './constructs/arRepository';
 
 export class CloudRunApp extends BaseConstruct {
   constructor(scope: Construct) {
     super(scope, 'CloudRunApp');
+
+    const serviceName = 'app';
+
+    const repo = new ArRepository(this, serviceName);
+
+    new CloudBuild(this, 'app', {
+      repo: repo.repo,
+      serviceName,
+      buildPermissions: ['secretmanager.versions.access'],
+      github: {
+        owner: 'cumet04',
+        name: 'switchbot-logger',
+        push: {
+          // mainが更新された場合は開発系も更新する
+          branch: this.env === 'production' ? '^main$' : `^(main|${this.env})$`,
+        },
+      },
+      buildYamlPath: 'app/cloudbuild.yaml',
+    });
 
     const token = new Secret(this, 'switchbot_token');
     const secret = new Secret(this, 'switchbot_secret');
@@ -23,9 +44,8 @@ export class CloudRunApp extends BaseConstruct {
       'secretmanager.versions.access',
     ]);
 
-    new CloudRun(this, 'app', {
+    new CloudRun(this, serviceName, {
       serviceAccount: sa,
-      buildPermissions: ['secretmanager.versions.access'],
       envvars: {
         PROJECT_ID: this.projectId.value,
         NEXT_PUBLIC_APP_ENV: this.env,
@@ -36,15 +56,6 @@ export class CloudRunApp extends BaseConstruct {
         SWITCHBOT_SECRET: secret,
         SENTRY_AUTH_TOKEN: sentryToken,
       },
-      github: {
-        owner: 'cumet04',
-        name: 'switchbot-logger',
-        push: {
-          // mainが更新された場合は開発系も更新する
-          branch: this.env === 'production' ? '^main$' : `^(main|${this.env})$`,
-        },
-      },
-      buildYamlPath: 'app/cloudbuild.yaml',
     });
   }
 }
